Lazy load the MV data
